Moneymaker/requirements.txt
# AI
anthropic>=0.30.0

# Web scraping / APIs
praw>=7.7.1
httpx>=0.27.0

# FastAPI control plane
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Celery task queue
celery[redis]>=5.4.0
flower>=2.0.1

# Databases
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
aiosqlite>=0.20.0          # dev/dry-run SQLite fallback

# Redis async client
redis[hiredis]>=5.0.4

# Observability
prometheus-client>=0.20.0

# Config / validation
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1

# Utilities
rich>=13.7.1
tenacity>=8.3.0

# Tests
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-mock>=3.14.0
